+++ src/components/MenuTab.jsx
@@ -1,18 +1,22 @@
 import "../style/header.css";
-import { BrowserRouter as Router, Link } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 
 function MenuTab() {
   return (
-    <Router>
-      <ul>
-        <li className="menu">자유게시판</li>
-        <li className="menu menu-active">
-          <Link to="/items" className="nav-link">
-            중고마켓
-          </Link>
-        </li>
-      </ul>
-    </Router>
+    <>
+      <NavLink
+        to="/board"
+        className={({ isActive }) => (isActive ? "menu menu-active" : "menu")}
+      >
+        자유게시판
+      </NavLink>
+      <NavLink
+        to="/items"
+        className={({ isActive }) => (isActive ? "menu menu-active" : "menu")}
+      >
+        중고마켓
+      </NavLink>
+    </>
   );
 }
