+++ src/pages/AddItemPage.js
@@ -1,0 +1,5 @@
+function AddItemPage() {
+  return <p>상품 등록하기</p>;
+}
+
+export default AddItemPage;
