+++ src/pages/BoardPage.js
@@ -1,0 +1,5 @@
+function BoardPage() {
+  return <p>자유게시판</p>;
+}
+
+export default BoardPage;
